- raised ac bullet damage from 6 to 7 - fixed sg lockon times using truncated integers (0) instead of their correct floats (0.5)

+++ dlls/ff/ff_sentrygun.cpp
@@ -100,11 +100,11 @@ ConVar sg_health_lvl3("ffdev_sg_health_lvl3", "200", FCVAR_REPLICATED, "Level 3 
 #define SG_HEALTH_LEVEL3	sg_health_lvl3.GetInt()
 
 ConVar sg_lockontime_lvl1("ffdev_sg_lockontime_lvl1", "1.00", FCVAR_REPLICATED, "Level 1 SG lock on time");
-#define SG_LOCKONTIME_LVL1	sg_lockontime_lvl1.GetInt()
+#define SG_LOCKONTIME_LVL1	sg_lockontime_lvl1.GetFloat()
 ConVar sg_lockontime_lvl2("ffdev_sg_lockontime_lvl2", "0.50", FCVAR_REPLICATED, "Level 2 SG lock on time");
-#define SG_LOCKONTIME_LVL2	sg_lockontime_lvl2.GetInt()
+#define SG_LOCKONTIME_LVL2	sg_lockontime_lvl2.GetFloat()
 ConVar sg_lockontime_lvl3("ffdev_sg_lockontime_lvl3", "0.50", FCVAR_REPLICATED, "Level 3 SG lock on time");
-#define SG_LOCKONTIME_LVL3	sg_lockontime_lvl3.GetInt()
+#define SG_LOCKONTIME_LVL3	sg_lockontime_lvl3.GetFloat()
 
 IMPLEMENT_SERVERCLASS_ST(CFFSentryGun, DT_FFSentryGun) 
 	SendPropInt( SENDINFO( m_iAmmoPercent), 8, SPROP_UNSIGNED ),
